+++ tests/genome/test_gff.py
@@ -2,7 +2,7 @@
 """
  * @Date: 2022-10-12 19:53:55
 * @LastEditors: hwrn hwrn.aou@sjtu.edu.cn
-* @LastEditTime: 2025-04-29 10:28:47
+* @LastEditTime: 2025-07-04 09:15:15
 * @FilePath: /genome/tests/genome/test_gff.py
  * @Description:
 __file__ = "tests/genome/test_prokka.py"
@@ -40,55 +40,6 @@ def test_gff_reset_reference(test_temp: Path):
         test_out,
         "fasta-2line",
     )
-
-
-def test_transl_except():
-    from Bio import SeqRecord, SeqFeature, Seq
-
-    # fmt: off
-    g1 = SeqRecord.SeqRecord(seq=Seq.Seq("ATGGTCCAATTCTCCAGACGCCAGTTCATGAAAGTGACTGGTTCGACTCTGGCGGGTTCCAGCTTGGCGCTGATGGGTTTCTCACCCACAGCTGCGCTTGCTGAGGTGCGACAGTACAAGCTGTCCGCCACCACCGTCACCCGCCAAACCTGCACCTACTGTTCCGTAGGCTGCGGCATCCTGATGTATTCCCTGGGCGACGGCGGCAAGAACAGCCGCCTGTCCGTGATGCACGTGGAAGGCGATCCGGATCATCCGGTCAATCGCGGCACGCTGTGTCCCAAGGGTGCATCGCTGCTGGACTTCGTGCACAGCCCCAGCCGTCTCAAATACCCCGAGTACCGCGCTCCCGGCTCCTCCGAATGGAAGCGCATGTCCTGGGACGAAGCGCTGGACCGCATCGCCAAGCTGCTCAAGGAAGACCGCGACGCCAATTTCGTTGAGACCAACGAAAAGGGCCAGAAGGTCAACCGCTGGCTCACCACCGGCATGCTGGCCGCTTCGGCCTCCAGCAACGAAGCCGGCTATATCACGCACAAAGTGGCCCGCTCCTGGGGCCTTCTTGCACTCGACAACCAAGCACGTGTCTGACACGGCCCGACGGTGGCAGGTCTTGCCCCGACGTTTGGCCGTGGAGCGATGACGAACCATTGGGTCGACATCAAGAACGCGGACGTTATTTTGATCATGGGCGGCAATGCCGCCGAAGCACACCC"),
-                             id="WP_309467740.1",
-                             features=[SeqFeature.SeqFeature(location=SeqFeature.SimpleLocation(195118, 195834, 1),
-                                                             qualifiers={"transl_except": ["(pos:195707..195709,aa:Sec)"]})])
-    assert gff.translate(g1, g1.features[0]).seq == "MVQFSRRQFMKVTGSTLAGSSLALMGFSPTAALAEVRQYKLSATTVTRQTCTYCSVGCGILMYSLGDGGKNSRLSVMHVEGDPDHPVNRGTLCPKGASLLDFVHSPSRLKYPEYRAPGSSEWKRMSWDEALDRIAKLLKEDRDANFVETNEKGQKVNRWLTTGMLAASASSNEAGYITHKVARSWGLLALDNQARVUHGPTVAGLAPTFGRGAMTNHWVDIKNADVILIMGGNAAEAH"
-    g2 = SeqRecord.SeqRecord(seq=Seq.Seq("CAGCCTTGGCGGAAGTACGCCAGTTCAAACTTACGCAGGCCAAGGTCACGCGCCAAACCTGCACGTACTGTTCTGTGGGCTGCGGCCTGCTGATGTATGCGCATGGCGATAGCGTGCGCAACGTCAAAGAAACCATCATGCACGTCGAAGGCGACCCGGATCACCCGGTCAATCGCGGCACGCTGTGCCCCAAGGGCGCGGCGCTGATGGACTTTGTCAATAGCCCCAACCGCCTGAAATTCCCCGAATACCGCGCCCCCGGCTCCAACGAGTGGAAGCGCATGAGCTGGGATGAGGCGCTGGACCGCATCGCCCGCCTGCTCAAAGACGACCGCGATGCCAACTTCCAAGAGCAAAACGCCTCTGGCCAGACGGTAAACCGCTGGTTGTCCACCGGCATGCTGGCCGCCTCGGCCTCCAGCAACGAAGCCGGTTACATCACGCACAAAGTGGCCCGCTCCTGGGGCCTGCTGGCACTCGATAACCAAGCACGTGTCTGACACGGCCCGACGGTGGCAGGTCTTGCCCCGACGTTTGGCCGTGGAGCCATGACAAACCATTGGGTGGACATCAAGAACGCAGATGTCATCCTCATCATGGGCGGCAATGCCGCAGAAGCGCACCCGTGCGGCTTCAAATGGGTGACCGAGGCCAAAGAGCACAACGGCGCTTATTTCATGGTGGTCGATCCACGCTTTAACCGCTCTGCGGCGGTCGCTGATTTTTATGCACCGATTCGCTCGGGCTCGGACATTGTGTTCCTGGGCGGCATGATCAATTACCTGCTCACGCACGACAAAATCCACCACGAGTATGTGAAGAACTACACGGACTTCAGCTTCATCGTGCGCGAGGATTTTGCGTTCGACGAGGGCATCTTCTCGGGCTACAACCCCGAGACGCGCACTTACGACAAATCCACCTGGGACTACGAGCTGGGCGAGGATGGCTTTGTCAAGGTCGATCCCACCTTGCAGCACCCGCGCTGCGTGTACCAGTGGCTCAAGCGCCACTACGCGGGCTACACCCCGGAAAAGGTCGAGAGCATCTGCGGCACGCCCAAGGAGAAATTCCTGCACGTGTGCGAGAAGTTTGCCTCCACCGCGCAAGCCGGTCGCGTGGCCACCATCATGTACGCCCTGGGCTGGACGCAGCACACCACGGGGGCGCAAATGCTGCGTACCGGTGCCTTGGTGCAGCTGCTGTGCGGCAACATCGGCGTGGCCGGCGGCGGCATGAATGCGCTGCGTGGGCACTCCAACATCCAGGGCCTGACCGATCTGGGCATTCTCTCGGCCTCGCTGCCGGGCTACCTGACCCTGCCCAACGAGAAGGAACAGGACTACGAGCAATTCATCGCCACGCGCACGCCCAAGCCGCTGCGCCCCGGCCAGATGAACTACTGGGCCAACTACTCCAAATTCCACGTCAGCCTGATGAAGGCCTGGTGGGGCCCAGCGGCCACGCAGGAGAACAATTGGGCGTTTGACTACCTGCCCAAGCTGGACCAGTCCTACGACATGCTCAAGATCTTTGACCTGATGAACCAGGGCAAGGTCAATGGCTACATCGCCCAGGGCTTCAACCCGCTGGCCTCGCTGGCCAACTCCAACAACGTGCGTGAGGGGCTGAAAAAGCTCAAGTTCCTCATCGTCATGGACCCGTTGGTGACCGAGACTTCCGAGTTCTGGAAAAACCACGGCGAGTACAACGATGTGGACCCCGCCAGCATCCAAACCGAGGTGTTCCGCCTGCCCACCACCTGCTTTGCCGAGGAAGATGGCGCCATCGTCAGCTCCTCGCGCGTGCTGCAGTGGCACTGGAAAGCGGCCGATGCCCCGGGTGAGGCCAAGACCGACATCACCATCATGTCGGCCCTGCACCTGCGCCTGAAAAAGCTCTACGCGCAAGAAGGCGGCAAGTTCCCCGACCCCATCACCAACCTGTACTGGCCCTATGCCCAGGCCGCGCACCCCAGCAGCGAAGAAATCGCCAAGGAGTACAACGGCCGTGCCCTGGTCGATCTGACCGATGCCGAGGGCAAAGTCATTCGCAAGGCAGGCGAGCAACTGGCCGGCTTTGGCGAGATGCGGGCCGATGGCAGCACCTTGGGCGGCTGCTGGATCTGGTCGGGCTGCTGGACCTCGTCCGGCAACCAGATGGCCCGGCGCGACAACAGCGACCCCACCGGCATTGGCAACACGCTGAACTGGGCCTGGGCCTGGCCGGCCAACCGCCGCGTGCTGTACAACCGCGCCTCGTGCGACCGCCAGGGCAAGCCCTTTGACCCGGATCGTGTGCTGGTGCAGTGGAACGGCAGCCGCTGGGTGGGCGCAGACGTGCCGGACGTGGGCCCCACGCTCGACCCCGAAACGGTGAACCCTTTCATCATGAACCCCGAAGGGGTGGCGCGCTTCTTCTCCCGCAAGGGCATGAACGAAGGGCCGTTCCCGACCCACTACGAGGCCTTTGACAACCCGCTGGGCTACAACCCGATGTACCCGGAGAACGAGCTGGCCGTCATCAGCCCGGCCGCACGGATTCTGGAATCGGCCAAGGGCACGGCGGGCGACCCCAAAGAGTTCCCGCACGTGGGCACCAGCTACCGCCTGACCGAGCACTTCCACTACTGGACCAAGCACGTGCACCTGAACAACATCGTGCAGCCGGAGCAGTTTGTGGAAATTGGCGAGGCGCTGGCCAAAGAGCTGGGCATCGAAACGGGCCAACAGGTCAAGGTCAGCTCCAAGCGCGGCTTTGTCAAGGCGGCGGCGGTCGTGACCAAGCGCATGAAGCCCATGCAGATCGACGGCAAGACCATCCACCACGTGGGCGTGCCCATCCACTGGGGCTTTGTCTCCCAGGGGCGCAAGGGGCACATGGCCAACAATCTGACGGCATCTGTGGGCGATGGCAACAGCTTCACCCCCGAATCCAAGACCTTCCTCGTGAAGGTCGAAAAGGTGTAA"),
-                             id="MBP6293671.1",
-                             features=[SeqFeature.SeqFeature(location=SeqFeature.SimpleLocation(0, 2969, 1),
-                                                             qualifiers={"transl_except": ["(pos:498..500,aa:Sec)"], "frame": ["2"]})])
-    assert gff.translate(g2, g2.features[0]).seq == "ALAEVRQFKLTQAKVTRQTCTYCSVGCGLLMYAHGDSVRNVKETIMHVEGDPDHPVNRGTLCPKGAALMDFVNSPNRLKFPEYRAPGSNEWKRMSWDEALDRIARLLKDDRDANFQEQNASGQTVNRWLSTGMLAASASSNEAGYITHKVARSWGLLALDNQARVUHGPTVAGLAPTFGRGAMTNHWVDIKNADVILIMGGNAAEAHPCGFKWVTEAKEHNGAYFMVVDPRFNRSAAVADFYAPIRSGSDIVFLGGMINYLLTHDKIHHEYVKNYTDFSFIVREDFAFDEGIFSGYNPETRTYDKSTWDYELGEDGFVKVDPTLQHPRCVYQWLKRHYAGYTPEKVESICGTPKEKFLHVCEKFASTAQAGRVATIMYALGWTQHTTGAQMLRTGALVQLLCGNIGVAGGGMNALRGHSNIQGLTDLGILSASLPGYLTLPNEKEQDYEQFIATRTPKPLRPGQMNYWANYSKFHVSLMKAWWGPAATQENNWAFDYLPKLDQSYDMLKIFDLMNQGKVNGYIAQGFNPLASLANSNNVREGLKKLKFLIVMDPLVTETSEFWKNHGEYNDVDPASIQTEVFRLPTTCFAEEDGAIVSSSRVLQWHWKAADAPGEAKTDITIMSALHLRLKKLYAQEGGKFPDPITNLYWPYAQAAHPSSEEIAKEYNGRALVDLTDAEGKVIRKAGEQLAGFGEMRADGSTLGGCWIWSGCWTSSGNQMARRDNSDPTGIGNTLNWAWAWPANRRVLYNRASCDRQGKPFDPDRVLVQWNGSRWVGADVPDVGPTLDPETVNPFIMNPEGVARFFSRKGMNEGPFPTHYEAFDNPLGYNPMYPENELAVISPAARILESAKGTAGDPKEFPHVGTSYRLTEHFHYWTKHVHLNNIVQPEQFVEIGEALAKELGIETGQQVKVSSKRGFVKAAAVVTKRMKPMQIDGKTIHHVGVPIHWGFVSQGRKGHMANNLTASVGDGNSFTPESKTFLVKVEKV*"
-    g3 = SeqRecord.SeqRecord(seq=Seq.Seq("ATGGTGCAATTCTCCAGACGCCAGTTCATGAAAGTGACTGGTTCGAGTCTGGCGGCATCTAGCTTGGCAGTGATGGGCTTCTCGCCCACTGCTGCCCTAGCAGAAGTTCGCCAGTTCAAGCTGGCAGCTGCGACAGTGACACGTCAAACGTGTACTTACTGTTCCGTAGGCTGCGGTATTTTGATGTATTCGTACAGCGACGGCGTAAAGAACGCCAAGATGTCCGTGATGCACGTTGAAGGTGATCCGGATCACCCAGTAAATCGCGGCACGCTGTGTCCCAAGGGTGCTTCGTTGCTCGATTTCGTGCACAGCCCCAACCGCCTTAAGTACCCCGAGTACCGTGCTCCTGGCTCTAACGAGTGGAAGCGTCTGTCTTGGGACGAAGCATTGAACCGCATCGCAAAACATCTGAAAGATGACCGCGACGCCAACTTCGTTGAAAAGAATGAAAAGGGCCAAACGGTTAACCGTTGGCTCACCACCGGCATGTTGGCAGCTTCTGCTTCCAGCAACGAAGCCGGTTATGTCACACATAAGGTAGCCCGTTCATGGGGTCTGCTGGCGCTCGATAACCAAGCACGTGTCTGACACGGCCCGACGGTGGCAGGTCTTGCCCCGACGTTTGGCCGTGGAGCCATGACCAATCACTGG"),
-                             id="MBF6632048.1",
-                             features=[SeqFeature.SeqFeature(location=SeqFeature.SimpleLocation(0, 654, -1),
-                                                             qualifiers={"transl_except": ["(pos:complement(64..66),aa:Sec)"]})])
-    assert gff.translate(g3, g3.features[0]).seq == "MVQFSRRQFMKVTGSSLAASSLAVMGFSPTAALAEVRQFKLAAATVTRQTCTYCSVGCGILMYSYSDGVKNAKMSVMHVEGDPDHPVNRGTLCPKGASLLDFVHSPNRLKYPEYRAPGSNEWKRLSWDEALNRIAKHLKDDRDANFVEKNEKGQTVNRWLTTGMLAASASSNEAGYVTHKVARSWGLLALDNQARVUHGPTVAGLAPTFGRGAMTNHW"
-    g4 = SeqRecord.SeqRecord(seq=Seq.Seq("GCTCGGTGGCCTGCGGCATCATCATGTACTCCAAGGGTGATTTTTACGCTGGTGAAAAGGCCAAGATCATCCACATCGAAGGCGATGCTGACCACCCCACCAACAAGGGCACGCTGTGCCCCAAGGGTGCGGCCATCAAGGACTTCGTCAACTCCGACACCCGCACCAAGGTGCCGCGCTACCGCGCACCGGGTGCAGACCATTGGGAAGAGGTCAGCTGGGACTTCGCCCTCGACAAGGTTGCGCGCCACATCAAGGATGACCGCGACGCCAACTTTGTCGAGAAGAACGAAGCCGGCACCACCGTGAACCGCTGGACCAGCATGGGTTTTCTGGCCGCCTCGGGCGGCACCAACGAAACCGGCTGGATCACCTACAAAACGGTGCGTGCCCTCGGCGTTCTGCCCTTCGACAATCAGGCACGCGTCTGACACGGCCCTACGGTAGCCAGTCTGGCTCCGACTTTTGGCCGTGGCGCAATGACCAACGCCTGGACCGATATCAAGAACACCGACCTCGTCGTCGTCATGGGTGGCAATGCCGCCGAAGCCCACCCCTGCGGCTTCAAATGGGTGACCGAGGCCAAACACCACAACAAGGCCAAGCTGATTGTTGTGGACCCCCGCTTCAACCGCACGGCATCGGTTGCGGATTACTACGCCCCCATACGGCCGGGCTCGGACATCTCCTTTCTGATGGGAGTGATCCGTTACTGCATGGCCAATGACAAGGTGCAGTGGGAATACCTCAAGCACTACACCAACGCCAGCTACCTGGTACGCGACGGCTTTGGCTGGAAGGATGGCCTTTTCAACGGCTATGACGAGGACAAGCGCGCCTACGACCAGAGCGAGTGGGACTACCAGTACGACGCCAGTGGCCAGCCGCTGACCGACCCCACGCTGCAACACCCGCGCTGCGTGTGGAATCTGCTGCGGGAACATGTAGCGGTGTACACGCCGGAGTTCGTGGAAAACATCTGCGGCACGCCCAAGGACCGTTTCCTGAAGATCTGCGAGATGATCGGAGAGACAGCGGCCACCGACAAGGTCATGACCTCCTGCTACGCCCTGGGCTGGACCCAGCATTCCAAGGGTGCGCAAAACATCCGCGGCATGGCCATGCTGCAGCTGATGCTGGGCAATATCGGCCTGCCCGGTGGCGGCGTGAACGCCTTGCGCGGCCACGCCAATATCCAGGGCCTGACCGACCTGGGTCTGCTCTCGCACATGGTGCCGGGCTATATGACCCTGCCCAACGAGCGTGAGACTTCGCTGCAGGCCTACCTGGCCACCAAGACCAACAAGCCCATGTTCCCGGGCCAGACCAGCTACTGGCAGCACACCGACAAGTTCATGGTGAGCTTTTTGAAGTCCATGTGGGGCGACAAGGCCACGGCAGAGAACGGTTTTGGCTATGACTGGCTGCCTAAGCTGGATGTGCCCTACTACGACATCCTGCGCACCTTCACCATGATGGAAGAGGGCAAGGTCAATCTGTACTTCTGCCAGGGTTTCAATCCGTTGATGTCTTTCCCCAACCGGGCCAAGCTGACGCGGGCGCTGTCCAAGCTCAAGCTGCTGGTGTCCATCGACCCGCTGGAGACCGAAACCGCTGCCTTCTGGCAAAACCACGGTGTCTACAACGATGTGAACCCGGCCGAGATCATGACCGAGGTGCTGGAGCTGCCCTCCACCTGCTTTGCCGAAGACGAAGGCGCCACGGTGAACTCCGGCCGCTGGCTACAGTGGCACTGGACCGGTGCCAACCCGCCCGGCAATGCCAAGCACGACACCTGGATCATGGCCGAGATCTTCCTGCGCGTGCGCGAGCTCTACCGCAAGGAAGGCGGCGTCTGCCCCGAGCCCATCTTCAACCTGAGCTGGGACTACCACAAGCCCAACGACCCCGAGCCCGAGGAACTGGCCAAGGAAATGAATGGCCGGGCGCTCAAGGACGTGTACGACCTCAACGACCCGAGCAGGCTGGTGCGCAAGGCCGGCGAGCAGCACGCGGCCTTCAGCGAGCTGCGCGCCGATGGCTCGACCATGTCGGGCTGCTGGATTTATGCCGGCTCCTGGACGGAGCAAGGAAACATGATGGCCCGCCGCAACAACCACGACGAAGGTGGCATAGGCACCTATCTGAACTGGACCTGGGCCTGGCCGGCCAACCGCCGCATTCTCTACAACCGCGCCAGCGCCGACCTCCAGGGCAAGCCCTGGGATGCCAAGCGCAAGCTGCTGGAGTGGAATGCGGACAAGCAGCGCTGGGAAGGCTTTGACGTGCCGGACTTTTCGCCCACGGCCCAGCCCGAGGTGGTGCGCCCCTTCATCATGACGCAGGAAGGCACGGCCCGGCTGTTTGTGCGCAAGATGCTGCGCGACGGCCCCTTCCCCACGCACATGGAGCCCTGGGAAAGCCCGGTGGAGAACCCCTTCAACAAGAAGATGCGCGGCAACCCCGTGGGCCGCATCTTCCCGGCCGACCTTGCACAGCTGGGCAGCAGCGAAGAGTTTCCTTTTGTAGGCACCTCGTACCGCCTGACCGAGCATTTCCACACCTGGACCAAGCACAACAAGGTGAACGCCGTGCTGCAGCCCGATTTCTTCGTCGAAATCAGTGAAGAGCTGGCGGCGCTGCGCGGCATTACCAAGGGTGGGCGCGTGCGCGTGTGGAACAAGCGCGGCGAGGTCTGGGCCAAGGCCGTGGTCACCAAGCGCATACGCCCGCTGATGTGCGACGGCAAGGAAGTGCATGTGATCGGCATCCCCATCCACTGGGGCTTCAAGGGCGCGGCCCGCAAGGGCTTTGGCGCCAACAGCCTGACGGTGTTTGCCGGTGAGGCGAATATCGACACGCCCGAGTCCAAGGCCTTTCTCGTCAACATCGAGCCGGCCAGCTCCGAAGACATCAAGAGGCTGACATCATGA"),
-                             id="MDR3005091.1",
-                             features=[SeqFeature.SeqFeature(location=SeqFeature.SimpleLocation(13035,15968, -1),
-                                                             qualifiers={"transl_except": ["(pos:complement(15538..15540),aa:Sec)"], "frame": ["2"]})])
-    assert gff.translate(g4, g4.features[0]).seq == "SVACGIIMYSKGDFYAGEKAKIIHIEGDADHPTNKGTLCPKGAAIKDFVNSDTRTKVPRYRAPGADHWEEVSWDFALDKVARHIKDDRDANFVEKNEAGTTVNRWTSMGFLAASGGTNETGWITYKTVRALGVLPFDNQARVUHGPTVASLAPTFGRGAMTNAWTDIKNTDLVVVMGGNAAEAHPCGFKWVTEAKHHNKAKLIVVDPRFNRTASVADYYAPIRPGSDISFLMGVIRYCMANDKVQWEYLKHYTNASYLVRDGFGWKDGLFNGYDEDKRAYDQSEWDYQYDASGQPLTDPTLQHPRCVWNLLREHVAVYTPEFVENICGTPKDRFLKICEMIGETAATDKVMTSCYALGWTQHSKGAQNIRGMAMLQLMLGNIGLPGGGVNALRGHANIQGLTDLGLLSHMVPGYMTLPNERETSLQAYLATKTNKPMFPGQTSYWQHTDKFMVSFLKSMWGDKATAENGFGYDWLPKLDVPYYDILRTFTMMEEGKVNLYFCQGFNPLMSFPNRAKLTRALSKLKLLVSIDPLETETAAFWQNHGVYNDVNPAEIMTEVLELPSTCFAEDEGATVNSGRWLQWHWTGANPPGNAKHDTWIMAEIFLRVRELYRKEGGVCPEPIFNLSWDYHKPNDPEPEELAKEMNGRALKDVYDLNDPSRLVRKAGEQHAAFSELRADGSTMSGCWIYAGSWTEQGNMMARRNNHDEGGIGTYLNWTWAWPANRRILYNRASADLQGKPWDAKRKLLEWNADKQRWEGFDVPDFSPTAQPEVVRPFIMTQEGTARLFVRKMLRDGPFPTHMEPWESPVENPFNKKMRGNPVGRIFPADLAQLGSSEEFPFVGTSYRLTEHFHTWTKHNKVNAVLQPDFFVEISEELAALRGITKGGRVRVWNKRGEVWAKAVVTKRIRPLMCDGKEVHVIGIPIHWGFKGAARKGFGANSLTVFAGEANIDTPESKAFLVNIEPASSEDIKRLTS*"
-    g5 = SeqRecord.SeqRecord(seq=Seq.Seq("AGCCTCCAGCCTGGCGGTACTGGGTTTTTCACCCAGCACAGCCTTGGCGGAAGTACGCCAGTTCAAACTTACGCAGGCCAAGGTCACGCGCCAAACCTGCACGTACTGTTCTGTGGGCTGCGGCCTGCTGATGTATGCGCATGGCGATAGCGTGCGCAACGTCAAAGAAACCATCATGCACGTCGAAGGCGACCCGGATCACCCGGTCAATCGCGGCACGCTGTGCCCCAAGGGCGCGGCGCTGATGGACTTTGTCAATAGCCCCAACCGCCTGAAATTCCCCGAATACCGCGCCCCCGGCTCCAACGAGTGGAAGCGCATGAGCTGGGATGAGGCGCTGGACCGCATCGCCCGCCTGCTCAAAGACGACCGCGATGCCAACTTCCAAGAGCAAAACGCCTCTGGCCAGACGGTAAACCGCTGGTTGTCCACCGGCATGCTGGCCGCCTCGGCCTCCAGCAACGAAGCCGGTTACATCACGCACAAAGTGGCCCGCTCCTGGGGCCTGCTGGCACTCGATAACCAAGCACGTGTCTGACACGGCCCGACGGTGGCAGGTCTTGCCCCGACGTTTGGCCGTGGAGCCATGACAAACCATTGGGTGGACATCAAGAACGCAGATGTCATCCTCATCATGGGCGGCAATGCCGCAGAAGCGCACCCGTGCGGCTTCAAATGGGTGACCGAGGCCAAAGAGCACAACGGCGCTTATTTCATGGTGGTCGATCCACGCTTTAACCGCTCTGCTGCGGTCGCTGATTTCTATGCACCGATTCGCTCGGGCTCGGACATTGTGTTCCTGGGCGGCATGATCAATTACCTGCTCACGCACGACAAAATCCACCACGAGTATGTGAAGAACTACACGGACTTCAGCTTCATCGTGCGCGAGGACTTTGCGTTCGACGAGGGCATCTTCTCGGGCTACAACCCCGAGACGCGCACTTACGACAAATCCACCTGGGACTACGAGCTGGGCGAAGATGGCTTTGTCAAGGTCGATCCCACCTTGCAGCACCCGCGCTGCGTGTACCAGTGGCTCAAGCGCCACTACGCGGGCTACACCCCGGAAAAGGTCGAGAGCATCTGCGGCACGCCCAAGGAGAAATTCCTGCACGTGTGCGAGAAGTTTGCCTCCACCGCGCAAGCCGGTCGCGTGGCCACCATCATGTACGCCCTGGGCTGGACGCAGCACACCACGGGGGCGCAAATGCTGCGTACCGGTGCCTTGGTGCAGCTGCTGTGCGGCAACATCGGCGTGGCCGGCGGCGGCATGAATGCGCTGCGTGGGCACTCCAACATCCAGGGGCTGACCGATCTGGGCATTCTCTCGGCCTCGCTGCCGGGCTACCTGACCCTGCCCAACGAGAAGGAACAGGACTACGAGCAATTCATCGCCACGCGCACGCCCAAGCCGCTGCGCCCCGGCCAGATGAACTACTGGGCCAACTACTCCAAATTCCACGTCAGCCTGATGAAGGCCTGGTGGGGCCCAGCGGCCACGCAGGAGAACAATTGGGCGTTTGACTACCTGCCCAAGCTGGACCAGTCCTACGACATGCTCAAGATCTTTGACCTGATGAACCAGGGCAAGGTCAATGGCTACATCGCCCAGGGCTTCAACCCCCTGGCCTCGCTGGCCAACTCCAACAACGTGCGTGAGGGGCTGAAAAAGCTCAAGTTCCTCATCGTCATGGACCCGTTGGTGACCGAGACTTCCGAGTTCTGGAAAAACCACGGCGAGTACAACGATGTGGACCCCGCCAGCATCCAAACCGAGGTGTTCCGCCTGCCCACCACCTGCTTTGCCGAGGAAGATGGCGCCATCGTCAGCTCCTCGCGCGTGCTGCAGTGGCACTGGAAAGCGGCCGATGCCCCGGGTGAGGCCAAGACCGACATCACCATCATGTCGGCCCTGCACCTGCGCCTGAAAAAGCTCTACGCGCAAGAAGGCGGCAAGTTCCCCGACCCCATCACCAACCTGTACTGGCCCTATGCCCAGGCCGCGCACCCCAGCAGCGAAGAAATCGCCAAGGAGTACAACGGCCGTGCCCTGGTCGATCTGACCGATGCCGAGGGCAAAGTCATTCGCAAGGCAGGCGAGCAACTGGCCGGCTTTGGCGAGATGCGGGCCGATGGCAGCACCTTGGGCGGCTGCTGGATCTGGTCGGGCTGCTGGACCTCGTCCGGCAACCAGATGGCCCGGCGCGACAACAGCGACCCCACCGGCATTGGCAACACGCTGAACTGGGCCTGGGCCTGGCCGGCCAACCGCCGCGTGCTGTACAACCGCGCCTCGTGCGACCGCCAGGGCAAGCCCTTTGACCCGGATCGTGTGCTGGTGCAGTGGAACGGCAGCCGCTGGGTGGGCGCAGACGTGCCGGACGTGGGCCCCACGCTCGACCCCGAAACGGTGAACCCTTTCATCATGAACCCCGAAGGGGTGGCGCGCTTCTTCTCCCGCAAGGGCATGAACGAAGGGCCGTTCCCGACCCACTACGAGGCCTTTGACAACCCGCTGGGCTACAACCCGATGTACCCGGAGAACAAGCTGGCCGTCATCAGCCCGGCCGCACGGATTCTGGAATCGGCCAAGGGCACGGCGGGCGACCCCAAAGAGTTCCCGCACGTGGGCACCAGCTACCGCCTGACCGAGCACTTCCACTACTGGACCAAGCACGTGCACCTGAACAACATCGTGCAGCCGGAGCAGTTTGTGGAAATTGGCGAGGCGCTGGCCAAAGAGCTGGGCATCGAAACGGGCCAACAGGTCAAGGTCAGCTCCAAGCGCGGCTTTGTCAAGGCGGCGGCGGTCGTGACCAAGCGCATGAAGCCCATGCAGATCGACGGCAAGACCATCCACCACGTGGGCGTGCCCATCCACTGGGGCTTTGTCTCCCAGGGGCGCAAGGGGCACATGGCCAACAATCTGACGGCATCTGTGGGCGATGGCAACAGCTTCACCCCCGAATCCAAGACCTTCCTCGTGAAGGTCGAAAAGGTGTAA"),
-                             id="MCZ2107622.1",
-                             features=[SeqFeature.SeqFeature(location=SeqFeature.SimpleLocation(6533, 9540, -1),
-                                                             qualifiers={"transl_except": ["(pos:complement(9003..9005),aa:Sec)"], "frame": ["1"]})])
-    assert gff.translate(g5, g5.features[0]).seq == "ASSLAVLGFSPSTALAEVRQFKLTQAKVTRQTCTYCSVGCGLLMYAHGDSVRNVKETIMHVEGDPDHPVNRGTLCPKGAALMDFVNSPNRLKFPEYRAPGSNEWKRMSWDEALDRIARLLKDDRDANFQEQNASGQTVNRWLSTGMLAASASSNEAGYITHKVARSWGLLALDNQARVUHGPTVAGLAPTFGRGAMTNHWVDIKNADVILIMGGNAAEAHPCGFKWVTEAKEHNGAYFMVVDPRFNRSAAVADFYAPIRSGSDIVFLGGMINYLLTHDKIHHEYVKNYTDFSFIVREDFAFDEGIFSGYNPETRTYDKSTWDYELGEDGFVKVDPTLQHPRCVYQWLKRHYAGYTPEKVESICGTPKEKFLHVCEKFASTAQAGRVATIMYALGWTQHTTGAQMLRTGALVQLLCGNIGVAGGGMNALRGHSNIQGLTDLGILSASLPGYLTLPNEKEQDYEQFIATRTPKPLRPGQMNYWANYSKFHVSLMKAWWGPAATQENNWAFDYLPKLDQSYDMLKIFDLMNQGKVNGYIAQGFNPLASLANSNNVREGLKKLKFLIVMDPLVTETSEFWKNHGEYNDVDPASIQTEVFRLPTTCFAEEDGAIVSSSRVLQWHWKAADAPGEAKTDITIMSALHLRLKKLYAQEGGKFPDPITNLYWPYAQAAHPSSEEIAKEYNGRALVDLTDAEGKVIRKAGEQLAGFGEMRADGSTLGGCWIWSGCWTSSGNQMARRDNSDPTGIGNTLNWAWAWPANRRVLYNRASCDRQGKPFDPDRVLVQWNGSRWVGADVPDVGPTLDPETVNPFIMNPEGVARFFSRKGMNEGPFPTHYEAFDNPLGYNPMYPENKLAVISPAARILESAKGTAGDPKEFPHVGTSYRLTEHFHYWTKHVHLNNIVQPEQFVEIGEALAKELGIETGQQVKVSSKRGFVKAAAVVTKRMKPMQIDGKTIHHVGVPIHWGFVSQGRKGHMANNLTASVGDGNSFTPESKTFLVKVEKV*"
-    # fmt: on
-
-    def report(rec: SeqRecord.SeqRecord):
-        """
-        e1 = _extract("GCF_031286875.1", "WP_309467740.1")
-        """
-        feat = rec.features[0]
-        frame = gff.check_frame(feat.qualifiers)
-        shift = frame > 0
-        loc = feat.location
-        assert loc
-        trans_except = gff.TranslExcept(feat.qualifiers["transl_except"][0])
-        pep_raw = gff.translate(rec, auto_fix=False).seq
-        if trans_except.strand == 1:
-            index = ((trans_except.start - 1) - (loc.start + shift)) // 3
-        else:
-            index = ((loc.end - shift) - (trans_except.end)) // 3
-        print(f"{rec.id} {frame} {loc} {trans_except} {pep_raw.index('*')} {index}")
 
 
 def test_to_dict():
